Fixes Bug #5: Hooks created in multiple children
Changed the install_hook method to handle the situation
where more then one child class defines a hook on a method
in the parent class.

Before this patch hooks in children subsiquent to the first
would not be called becuase the mutated method (hook.rb:274f)
placed in the first sibling.

This patch moves the hook stack exicution method to the common
ancestor of the two classes, and makes an addational call to
define_advize_method.