Another Edge case for inheritance Fixed
Please see changed to spec/hook_spec.rb for details.

Synopsis: The Grnadchild of an originlly hooked child
would not recieve its hooks when that child's hooks
were moved to accomodate it's sibling.

We now remove the original hook stack execution methods
in favor of those on the parent.

It seems like there ought to be a better way to extract
these, but his is working.